+++ src/Foundation/Bootstrap/DetectDomain.php
@@ -4,21 +4,22 @@ use Illuminate\Contracts\Foundation\Application;
 
 class DetectDomain {
 
-	/**
-	 * Bootstrap the given application.
-	 *
-	 * @param  \Illuminate\Contracts\Foundation\Application  $app
-	 * @return void
-	 */
-	public function bootstrap(Application $app)
-	{
+    /**
+     * Bootstrap the given application.
+     *
+     * @param  \Illuminate\Contracts\Foundation\Application  $app
+     * @return void
+     */
+    public function bootstrap(Application $app)
+    {
+        $app->detectDomain();
 
-        //Detect the domain
-		$app->detectDomain();
+        $storagePath = $app->domainStoragePath();
+        $defaultStoragePath = rtrim($app->basePath('storage'), '\/');
 
-        //Overrides the storage path if the domain stoarge path exists
-        //$app->useStoragePath($app->domainStoragePath());
-
-	}
+        if ($storagePath !== $defaultStoragePath) {
+            $app->useStoragePath($storagePath);
+        }
+    }
 
 }