+++ src/Foundation/Console/Kernel.php
@@ -17,6 +17,7 @@ class Kernel extends \Illuminate\Foundation\Console\Kernel {
         \Gecche\Multidomain\Foundation\Bootstrap\DetectDomain::class,
         \Illuminate\Foundation\Bootstrap\LoadEnvironmentVariables::class,
         \Illuminate\Foundation\Bootstrap\LoadConfiguration::class,
+        \Gecche\Multidomain\Foundation\Bootstrap\LoadDomainConfiguration::class,
         \Illuminate\Foundation\Bootstrap\HandleExceptions::class,
         \Illuminate\Foundation\Bootstrap\RegisterFacades::class,
         \Illuminate\Foundation\Bootstrap\SetRequestForConsole::class,