+++ src/Foundation/Http/Kernel.php
@@ -18,6 +18,7 @@ class Kernel extends \Illuminate\Foundation\Http\Kernel {
         \Gecche\Multidomain\Foundation\Bootstrap\DetectDomain::class,
         \Illuminate\Foundation\Bootstrap\LoadEnvironmentVariables::class,
         \Illuminate\Foundation\Bootstrap\LoadConfiguration::class,
+        \Gecche\Multidomain\Foundation\Bootstrap\LoadDomainConfiguration::class,
         \Illuminate\Foundation\Bootstrap\HandleExceptions::class,
         \Illuminate\Foundation\Bootstrap\RegisterFacades::class,
         \Illuminate\Foundation\Bootstrap\RegisterProviders::class,