+++ tests/app/Console/Kernel.php
@@ -2,6 +2,7 @@
 
 namespace Gecche\Multidomain\Tests\App\Console;
 
+use Gecche\Multidomain\Tests\App\Console\Commands\ConfigNameCommand;
 use Gecche\Multidomain\Tests\App\Console\Commands\NameCommand;
 use Gecche\Multidomain\Tests\App\Console\Commands\QueuePushCommand;
 use Gecche\Multidomain\Foundation\Console\Kernel as ConsoleKernel;
@@ -16,6 +17,7 @@ class Kernel extends ConsoleKernel
      */
     protected $commands = [
         NameCommand::class,
+        ConfigNameCommand::class,
         QueuePushCommand::class,
         //
     ];